refactor(interfaces): renamed interfaces to follow single pattern, created a generic crud interface to centralize the approach

+++ apps/backend/src/user/user.service.ts
@@ -7,11 +7,11 @@ import { saltRounds } from '../common/constants';
 import { Repository } from 'typeorm';
 import { User } from './entities/user.entity';
 import { v4 as uuidv4 } from 'uuid';
-import { UserServiceInterface } from './interfaces/user-service.interface';
+import { IUserService } from './interfaces/user-service.interface';
 import * as merge from 'lodash.merge';
 
 @Injectable()
-export class UserService implements UserServiceInterface {
+export class UserService implements IUserService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
@@ -51,7 +51,7 @@ export class UserService implements UserServiceInterface {
   }
 
   async findById(id: string): Promise<User> {
-    if(!id) throw new NotFoundException(`User with id ${id} not found`);
+    if (!id) throw new NotFoundException(`User with id ${id} not found`);
     const user: User = await this.userRepository.findOne({ where: { id } });
     if (!user) throw new NotFoundException(`User with id ${id} not found`);
     return user;
@@ -74,6 +74,7 @@ export class UserService implements UserServiceInterface {
 
   async remove(id: string): Promise<User> {
     const foundUser: User = await this.findById(id);
+    if (!foundUser) throw new NotFoundException(`User with id ${id} not found`);
     return await this.userRepository.remove(foundUser);
   }
 }
